fix: удалил pragma disable, переписал Settings.cs, чуток переписал, чтобы красиво было

+++ SolarObjects/ISolarObject.cs
@@ -5,7 +5,7 @@ namespace SolarObjects;
 public interface ISolarObject
 {
     Vector2 Coordinates { get; }
-    int Mass { get; }
+    float Mass { get; }
     void InteractWithAnotherObject(ISolarObject solarObject);
     void Update();
 }
